FastSAM_TensorRT/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastsam-trt"
version = "0.1.0"
description = "TensorRT-accelerated FastSAM inference"
requires-python = ">=3.8"
dependencies = [
    "numpy",
    "opencv-python",
    "torch",
    "ultralytics",
    "onnx",
    "onnxsim",
    "onnxruntime-gpu",
    "tensorrt",
    "pycuda",
]

[tool.setuptools.packages.find]
where = ["src"]
